+++ NEWS.md
@@ -2,6 +2,14 @@
 
 -   Fix geometries on `esp_get_hex_prov()` and `esp_get_hex_ccaa()`.
 
+-   Add new function to get comarcas from INE: `esp_get_comarca()`.
+
+-   Add new functions to get simplified maps from INE:
+
+    -   `esp_get_simpl_prov()`
+
+    -   `esp_get_simpl_ccaa()`
+
 # mapSpain 0.6.1
 
 -   HOTFIX: Bug on `esp_getTiles()` when `mask = TRUE`.